refactor SessionContext
