+++ src/components/Layout/Navigation/Sidebar.jsx
@@ -7,7 +7,7 @@ import useScreenType from '../../../hooks/useScreenType';
 import Logo from '../../Logo/Logo';
 import { Trans, useTranslation } from 'react-i18next';
 import StatusContext from '../../../context/StatusContext';
-import SessionContext from '../../../context/SessionContext';
+import SessionContext from '@/context/SessionContext';
 import { MdNotifications } from "react-icons/md";
 import ConnectionStatusIcon from './ConnectionStatusIcon';
