+++ src/components/Popups/SelectCredentialsPopup.jsx
@@ -5,7 +5,7 @@ import { useTranslation, Trans } from 'react-i18next';
 import CredentialImage from '../Credentials/CredentialImage';
 import CredentialInfo from '../Credentials/CredentialInfo';
 import Button from '../Buttons/Button';
-import SessionContext from '../../context/SessionContext';
+import SessionContext from '@/context/SessionContext';
 import useScreenType from '../../hooks/useScreenType';
 import Slider from '../Shared/Slider';
 import CredentialParserContext from '@/context/CredentialParserContext';
