+++ src/lib/services/OpenID4VCI/OpenID4VCI.tsx
@@ -9,7 +9,7 @@ import * as config from '../../../config';
 import { useHttpProxy } from '../HttpProxy/HttpProxy';
 import { useOpenID4VCIClientStateRepository } from '../OpenID4VCIClientStateRepository';
 import { useCallback, useContext, useMemo, useEffect, useRef } from 'react';
-import SessionContext from '../../../context/SessionContext';
+import SessionContext from '@/context/SessionContext';
 import CredentialsContext from '@/context/CredentialsContext';
 import { useOpenID4VCIPushedAuthorizationRequest } from './OpenID4VCIAuthorizationRequest/OpenID4VCIPushedAuthorizationRequest';
 import { useOpenID4VCIAuthorizationRequestForFirstPartyApplications } from './OpenID4VCIAuthorizationRequest/OpenID4VCIAuthorizationRequestForFirstPartyApplications';
