+++ src/pages/AddCredentials/AddCredentials.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect, useContext } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import StatusContext from '../../context/StatusContext';
-import SessionContext from '../../context/SessionContext';
+import SessionContext from '@/context/SessionContext';
 import RedirectPopup from '../../components/Popups/RedirectPopup';
 import { H1 } from '../../components/Shared/Heading';
 import PageDescription from '../../components/Shared/PageDescription';
