+++ src/pages/Home/Home.jsx
@@ -4,7 +4,7 @@ import { useNavigate } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
 
 // Contexts
-import SessionContext from '../../context/SessionContext';
+import SessionContext from '@/context/SessionContext';
 import CredentialsContext from '@/context/CredentialsContext';
 
 // Hooks
