+++ src/pages/Login/LoginState.jsx
@@ -5,7 +5,7 @@ import { GoPasskeyFill } from 'react-icons/go';
 import { Trans, useTranslation } from 'react-i18next';
 
 import StatusContext from '../../context/StatusContext';
-import SessionContext from '../../context/SessionContext';
+import SessionContext from '@/context/SessionContext';
 
 import LanguageSelector from '../../components/LanguageSelector/LanguageSelector';
 import Button from '../../components/Buttons/Button';
