+++ src/pages/Settings/Settings.tsx
@@ -6,7 +6,7 @@ import { MdNotifications } from "react-icons/md";
 import { IoIosArrowDown } from "react-icons/io";
 
 import StatusContext from '../../context/StatusContext';
-import SessionContext from '../../context/SessionContext';
+import SessionContext from '@/context/SessionContext';
 
 import useScreenType from '../../hooks/useScreenType';
